### 2026-01-12 #102 @andreluza
[¬Re] Bias correction in species distribution models: pooling survey and collection data for multiple species
Dear @ReScience/editors,

I request a review for the following replication:

### Original article

**Title:** Bias correction in species distribution models: pooling survey and collection data for multiple species
**Author(s):**   Fithian, W., Elith, J., Hastie, T. and Keith, D.A.
**Journal (or Conference):**  Methods Ecol Evol, 6: 424-438.
**Year:**  2015
**DOI:**  10.1111/2041-210X.12242
**PDF:**   https://besjournals.onlinelibrary.wiley.com/doi/full/10.1111/2041-210x.12242

### Replication

**Author(s)**:   Luza, A.L., Barraquand, F.
**Repository**:  https://github.com/andreluza/ISDM_replication/tree/main
**PDF**:  https://github.com/andreluza/ISDM_replication/blob/main/Article/article.pdf
**Keywords**:  Integrated Species Distribution Models, Multi-species models, Poisson Point Process, Sampling bias.
**Language**:  R
**Domain**:  Ecology

### Results

* [ ] Article has been fully replicated
* [ ] Article has been partially replicated
* [X] Article has not been replicated

### Potential reviewers
@tpoisot
**Suggested reviewers:**
  Philipp Boersch-Supan and Daniel Flynn

### Invited reviewers

- [ ] @sampeel
- [ ] @PhilipMostert
