### 2026-02-05 #102 @tpoisot
@rougier I would gladly handle this one as editor
