### 2026-03-10 #102 @tpoisot
@andreluza can you come up with a shortlist of ~5 reviewers? I'll add a few names of my own, but it's difficult enough to find reviewers for regular papers, so code is usually more difficult...
