### 2026-03-13 #102 @andreluza
Sure @tpoisot . Below is a list of six names (with a link to their GitHub page, if needed).

Philip Mostert: https://github.com/PhilipMostert
Bob O'Hara: https://github.com/oharar
Samantha L Peel: https://github.com/sampeel
Nick Golding: https://github.com/goldingn
Ben Goldstein: https://github.com/dochvam
Viviane Zulian: https://github.com/vivizulian
