### 2026-03-13 #102 @tpoisot
Thank you @andreluza

@sampeel and @PhilipMostert - would you be available to review this replication paper? Information about the review process can be found here: https://rescience.github.io/edit/
